fix unsigned->size_t build regression

+++ src/lzma-encode.cpp
@@ -27,8 +27,8 @@ namespace vm {
 uint8_t* encodeLZMA(System* s,
                     avian::util::Alloc* a,
                     uint8_t* in,
-                    unsigned inSize,
-                    unsigned* outSize)
+                    size_t inSize,
+                    size_t* outSize)
 {
   const unsigned PropHeaderSize = 5;
   const unsigned HeaderSize = 13;
